# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SynthChat Hermes Backend API
  version: 0.1.0
  description: |
    Frontend-facing contract for the complete Rust backend. Upstream Hermes
    HTTP, dashboard, WebSocket, CLI, filesystem and SQLite interfaces are
    compatibility references only and are not production runtime dependencies.
servers:
  - url: http://127.0.0.1:8642
    description: Local desktop backend
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/providers:
    get:
      tags: [Profiles]
      operationId: listProviders
      summary: List provider catalog entries
      responses:
        "200":
          description: Provider catalog without credentials.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Provider" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/profiles:
    get:
      tags: [Profiles]
      operationId: listProfiles
      summary: List profiles
      responses:
        "200":
          description: Profiles visible to this local installation.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Profiles]
      operationId: createProfile
      summary: Create a profile
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProfile" }
      responses:
        "201":
          description: Profile created.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileMetadata" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/profiles/{profileId}:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    get:
      tags: [Profiles]
      operationId: getProfile
      summary: Get profile metadata
      responses:
        "200":
          description: Profile metadata.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileMetadata" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Profiles]
      operationId: updateProfile
      summary: Update profile metadata
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/ProfilePatch" }
      responses:
        "200":
          description: Updated profile.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileMetadata" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
    delete:
      tags: [Profiles]
      operationId: deleteProfile
      summary: Delete a profile
      responses:
        "204": { description: Profile deleted or already absent. The default and active profiles cannot be deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/profiles/{profileId}/active:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    put:
      tags: [Profiles]
      operationId: activateProfile
      summary: Activate a profile
      responses:
        "200":
          description: Active profile. Existing sessions and runs remain pinned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/profiles/{profileId}/config:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    get:
      tags: [Profiles]
      operationId: getProfileConfig
      summary: Get profile configuration
      responses:
        "200":
          description: Non-secret structured Hermes configuration.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Profiles]
      operationId: updateProfileConfig
      summary: Update profile configuration
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/ProfileConfigPatch" }
      responses:
        "200":
          description: Updated non-secret configuration. Unknown YAML keys are preserved.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileConfig" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
...
  /api/v1/profiles/{profileId}/secrets:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    get:
      tags: [Profiles]
      operationId: listSecretStatuses
      summary: List keychain secret statuses
      responses:
        "200":
          description: All Rust model and Web provider-catalog secret names plus indexed custom names. Values, previews and lengths are never returned.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SecretStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/profiles/{profileId}/secrets/{secretName}:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
      - $ref: "#/components/parameters/SecretName"
    put:
      tags: [Profiles]
      operationId: putSecret
      summary: Store a secret in the OS keychain
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SecretValue" }
      responses:
        "200":
          description: Secret stored in the OS keychain.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecretStatus" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    delete:
      tags: [Profiles]
      operationId: deleteSecret
      summary: Delete a keychain secret
      responses:
        "204": { description: Secret removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /api/v1/files:
    post:
      tags: [Files]
      operationId: uploadFile
      summary: Upload a file
      description: Stores one opaque file snapshot under HERMES_HOME without accepting a filesystem path. The semantic request fingerprint covers the validated filename, normalized MIME type and complete bytes.
      x-max-file-bytes: 8388608
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: File accepted and represented by an opaque ID.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileRef" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "507": { $ref: "#/components/responses/InsufficientStorage" }
...
  /api/v1/files/{fileId}:
    parameters:
      - $ref: "#/components/parameters/FileId"
    delete:
      tags: [Files]
      operationId: deleteFile
      summary: Delete a temporary file
      responses:
        "204": { description: File deleted or already absent. A retained idempotency record is not deleted and will never recreate an explicitly deleted file. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-desktop-session-token
  headers:
    ETag:
      description: Opaque resource revision, including quotes as required by HTTP.
      schema: { type: string }
  parameters:
    ProfileId: { name: profileId, in: path, required: true, schema: { $ref: "#/components/schemas/ProfileIdValue" } }
...
    FileId: { name: fileId, in: path, required: true, schema: { type: string, pattern: '^file_[0-9a-f]{32}$' } }
...
    SecretName: { name: secretName, in: path, required: true, schema: { $ref: "#/components/schemas/SecretName" } }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: A single quoted strong ETag. Weak, wildcard, multiple or malformed values are invalid.
      schema: { type: string }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required for non-idempotent creates. Scoped by installation, method and canonical path; retained for at least 24 hours. Reuse with a different request fingerprint returns 409.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    BadRequest:
      description: Invalid request.
      x-error-codes: [validation_failed, invalid_if_match, invalid_cursor]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Missing or invalid desktop token.
      x-error-codes: [unauthorized]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found.
      x-error-codes: [resource_not_found, profile_not_found, approval_not_found, clarification_not_found, hermes_state_not_found]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: State, revision, idempotency, replay-window or busy conflict.
      x-error-codes: [revision_conflict, idempotency_conflict, idempotency_resource_gone, mcp_server_exists, mcp_config_invalid, capacity_exceeded, unsafe_file_path, memory_storage_drift, profile_delete_conflict, session_busy, session_archived, event_history_expired, hermes_import_source_changed, hermes_import_conflict]
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    PreconditionRequired:
      description: A single strong If-Match value is required.
      x-error-codes: [precondition_required]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ServiceUnavailable:
      description: A required local service or protected store is unavailable.
      x-error-codes: [secret_storage_unavailable, file_storage_unavailable, skill_storage_unavailable, skill_management_unavailable, operation_storage_unavailable]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InsufficientStorage:
      description: A bounded local retained-object quota is exhausted; callers must delete no-longer-needed objects before retrying.
      x-error-codes: [file_quota_exceeded]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    PayloadTooLarge:
      description: Upload or body exceeds the advertised limit.
      x-error-codes: [payload_too_large, hermes_import_too_large]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnsupportedMediaType:
      description: The request does not carry the single required media type.
      x-error-codes: [unsupported_media_type, unsupported_file_type]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Provider:
      type: object
      additionalProperties: false
      required: [id, displayName, defaultBaseUrl, requiresSecret, secretNames, supportsModelDiscovery]
      properties:
        id: { type: string }
        displayName: { type: string }
        defaultBaseUrl:
          type: [string, "null"]
          format: uri
          pattern: "^[Hh][Tt][Tt][Pp][Ss]?://"
          description: When non-null, requires a host and forbids userinfo, query and fragment components.
        requiresSecret: { type: boolean }
        secretNames: { type: array, items: { $ref: "#/components/schemas/SecretName" } }
        supportsModelDiscovery: { type: boolean }
...
    Profile:
      type: object
      additionalProperties: false
      required: [id, displayName, isDefault, isActive, engineState, configRevision, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/ProfileIdValue" }
        displayName: { $ref: "#/components/schemas/ProfileDisplayName" }
        isDefault: { type: boolean }
        isActive: { type: boolean }
        color: { $ref: "#/components/schemas/ProfileColor" }
        avatarFileId: { type: [string, "null"] }
        engineState:
          type: string
          enum: [stopped, starting, running, degraded, failed]
          description: Shared on-demand Rust Run service readiness for this profile; it is not a Provider credential or endpoint preflight result.
        configRevision: { type: string }
        createdAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
    ProfileMetadata:
      type: object
      additionalProperties: false
      required: [id, displayName, isDefault, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/ProfileIdValue" }
        displayName: { $ref: "#/components/schemas/ProfileDisplayName" }
        isDefault: { type: boolean }
        color: { $ref: "#/components/schemas/ProfileColor" }
        avatarFileId: { type: [string, "null"] }
        createdAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
    ProfileIdValue:
      type: string
      pattern: "^(default|[a-z0-9_][a-z0-9_-]{0,63})$"
    NamedProfileId:
      allOf:
        - $ref: "#/components/schemas/ProfileIdValue"
      not: { const: default }
    ProfileDisplayName:
      type: string
      minLength: 1
      maxLength: 80
    ProfileColor:
      type: [string, "null"]
      pattern: "^#[0-9A-Fa-f]{6}$"
    CreateProfile:
      type: object
      additionalProperties: false
      required: [id, displayName]
      properties:
        id: { $ref: "#/components/schemas/NamedProfileId" }
        displayName: { $ref: "#/components/schemas/ProfileDisplayName" }
        cloneFromProfileId:
          oneOf:
            - $ref: "#/components/schemas/ProfileIdValue"
            - type: "null"
    ProfilePatch:
      type: object
      additionalProperties: false
      description: Constrained merge patch. Null clears color or avatarFileId and is rejected for every other field.
      x-merge-patch-profile:
        explicitNullPaths: [/color, /avatarFileId]
        rejectOtherNulls: true
      properties:
        displayName: { $ref: "#/components/schemas/ProfileDisplayName" }
        color:
          allOf:
            - $ref: "#/components/schemas/ProfileColor"
          description: Null clears the stored color.
        avatarFileId:
          type: [string, "null"]
          description: Null clears the stored avatar.
    ModelConfig:
      type: object
      additionalProperties: false
      required: [provider, model, baseUrl]
      properties:
        provider: { type: string }
        model: { type: string }
        baseUrl:
          type: [string, "null"]
          format: uri
          pattern: "^[Hh][Tt][Tt][Pp][Ss]?://"
          description: Null selects the provider default. Non-null values require a host and forbid userinfo, query and fragment components.
        reasoningEffort: { type: [string, "null"], enum: [minimal, low, medium, high, xhigh, null] }
    ProfileConfig:
      type: object
      additionalProperties: false
      required: [revision, model, codeExecution, toolsets, skills, memoryProvider, platforms, extensions]
      properties:
        revision: { type: string }
        model: { $ref: "#/components/schemas/ModelConfig" }
        codeExecution: { $ref: "#/components/schemas/CodeExecutionConfig" }
        toolsets:
          type: object
          description: Normalized CLI toolset state backed by platform_toolsets.cli and agent.disabled_toolsets. Disabled entries override enabled entries.
          additionalProperties: { type: boolean }
        skills:
          type: object
          readOnly: true
          description: Derived enablement projection from the dedicated Skills directory and nested configuration subsystem. It may be empty until that subsystem is available and is never writable through ProfileConfigPatch.
          additionalProperties: { type: boolean }
        memoryProvider: { type: string }
        platforms:
          type: object
          readOnly: true
          description: Derived platform availability from keychain/environment state plus top-level platform enabled=false overrides. It may be empty until dedicated platform endpoints are available and is never writable through ProfileConfigPatch.
          additionalProperties: { type: boolean }
        extensions:
          type: object
          additionalProperties: { $ref: "#/components/schemas/NonNullJsonValue" }
    ProfileConfigPatch:
      type: object
      additionalProperties: false
      description: Constrained merge patch for model, codeExecution, toolsets, memoryProvider and extensions only. Skills and platforms are derived by dedicated subsystems. Null is an explicit state only for model.baseUrl and model.reasoningEffort; all other nulls are rejected and never delete properties.
      x-merge-patch-profile:
        explicitNullPaths: [/model/baseUrl, /model/reasoningEffort]
        nullDeletesProperties: false
        rejectOtherNulls: true
      properties:
        model: { $ref: "#/components/schemas/ModelConfigPatch" }
        codeExecution: { $ref: "#/components/schemas/CodeExecutionConfigPatch" }
        toolsets:
          type: object
          description: Field-level desired state. True inserts the ID into platform_toolsets.cli and removes it from agent.disabled_toolsets; false removes it from platform_toolsets.cli and inserts it into agent.disabled_toolsets.
          additionalProperties: { type: boolean }
        memoryProvider: { type: string }
        extensions:
          type: object
          description: Non-secret extension values. Null is rejected at every nesting level.
          additionalProperties: { $ref: "#/components/schemas/NonNullJsonValue" }
...
    CodeExecutionConfigPatch:
      type: object
      additionalProperties: false
      properties:
        mode: { type: string, enum: [project, strict] }
        timeoutSeconds: { type: integer, minimum: 1, maximum: 600 }
        maxToolCalls: { type: integer, minimum: 1, maximum: 100 }
...
    SecretStatus:
      type: object
      additionalProperties: false
      required: [name, configured, storage]
      properties:
        name: { $ref: "#/components/schemas/SecretName" }
        configured: { type: boolean }
        storage: { type: string, const: osKeychain }
        updatedAt: { type: [string, "null"], format: date-time }
    SecretName:
      type: string
      pattern: "^[A-Z][A-Z0-9_]{0,127}$"
    SecretValue:
      type: object
      additionalProperties: false
      required: [value]
      properties:
        value: { type: string, minLength: 1, maxLength: 2560, writeOnly: true }
...
    FileRef:
      type: object
      additionalProperties: false
      required: [id, name, mimeType, sizeBytes, createdAt]
      properties:
        id: { type: string, pattern: '^file_[0-9a-f]{32}$' }
        name: { type: string, minLength: 1, maxLength: 255 }
        mimeType: { type: string, enum: [application/json, application/octet-stream, application/pdf, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/x-zip-compressed, application/yaml, application/zip, image/gif, image/jpeg, image/png, image/webp, text/csv, text/markdown, text/plain, text/tab-separated-values, text/yaml] }
        sizeBytes: { type: integer, minimum: 0, maximum: 8388608 }
        createdAt: { type: string, format: date-time }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId, retryable]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: [string, "null"] }
        instance: { type: [string, "null"], format: uri-reference }
        code:
          type: string
          description: Stable machine-readable code. Each response component's x-error-codes lists known values without closing the set to future additions.
        requestId: { type: string }
        retryable: { type: boolean }